Refactor SIMD logic

+++ src/nn/embedded_weight.hpp
@@ -1,4 +1,4 @@
 namespace nn {
   extern const char* kEmbeddedWeight;
-  extern const size_t kEmbeddedWeightSize;
+  extern const int kEmbeddedWeightSize;
 };
